Lots of changes thanks to some suggestions from @joohoi -------------------------------------------------------- Pattern had .* surrounding it to show the context of that match. This is no longer the case. Now searches are done line by line allowing for more powerful regex searches like case insensitivity and ^ $ delimiters.
To enable context there is a new flag "context" which by default shows
50 characters before and after the match or the beginning and ending of
the line which ever is smaller. This allows searching more cleanly in
minified JS.

The output for matches is now a bit more verbose telling you which line
and the offset within that line that the match was found.